+++ fpsim/sim.py
@@ -100,8 +100,10 @@ class People(fpb.BasePeople):
         self.partnership_age = arr(n, d['partnership_age'])     # Age at first partnership in years, initialised from data
         self.urban           = arr(n, d['urban'])               # Whether a person lives in rural or urban setting
         self.paid_employment = arr(n, d['paid_employment'])     # Whether a person has a paid job or not
-        self.control_over_wages = arr(n, d['control_over_wages'])   # Decision making autonomy over major household purchases
-        self.sexual_autonomy    = arr(n, d['sexual_autonomy'])      # Ability to refuse sex
+        self.decision_wages  = arr(n, d['decision_wages'])       # Decision making autonomy over major household purchases/wages
+        self.decision_health = arr(n, d['decision_health'])      # Decision making autonomy over her health
+        self.sexual_autonomy    = arr(n, d['sexual_autonomy'])   # Ability to refuse sex
+
         # Empowerment-education attributes
         self.edu_objective   = arr(n, d['edu_objective'])   # Highest-ideal level of education to be completed (in years), could be individualised or constant across agents
         self.edu_attainment  = arr(n, d['edu_attainment'])  # Current level of education achieved in years
@@ -1272,7 +1274,9 @@ class Sim(fpb.BaseSim):
         empowerment = {}
         empowerment['paid_employment'] = np.zeros(n, dtype=bool)
         empowerment['sexual_autonomy'] = np.zeros(n, dtype=float)
-        empowerment['control_over_wages'] = np.zeros(n, dtype=float)
+        empowerment['decision_wages'] = np.zeros(n, dtype=float)
+        empowerment['decision_health'] = np.zeros(n, dtype=float)
+
 
         if empowerment_dict is not None:
             # Find only female agents
@@ -1289,7 +1293,7 @@ class Sim(fpb.BaseSim):
             paid_employment_probs = empowerment_dict['paid_employment']
             empowerment['paid_employment'][f_inds] = fpu.binomial_arr(paid_employment_probs[age_inds])
 
-            for metric in ['control_over_wages', 'sexual_autonomy']:
+            for metric in ['decision_wages', 'decision_health', 'sexual_autonomy']:
                 empowerment[metric][f_inds] = empowerment_dict[metric][age_inds]
         return empowerment
 
@@ -1506,7 +1510,8 @@ class Sim(fpb.BaseSim):
             partnership_age=p.partnership_age,
             paid_employment=p.paid_employment,
             sexual_autonomy=p.sexual_autonomy,
-            control_over_wages=p.control_over_wages,
+            decision_wages=p.decision_wages,
+            decision_health=p.decision_health,
             edu_objective=p.edu_objective,
             edu_attainment=p.edu_attainment,
             edu_completed=p.edu_completed,